Split CategoriesFavoritesSegmentControl into separate Preview file, fix preview
- Add assignment of configuration var, this was failing the assertion
- Moving to a separate file fixed the other errors, see:
> SwiftUI preview: error: '__designTimeInteger(_:fallback:)' is only available in iOS 13.0 or newer - https://forums.developer.apple.com/forums/thread/716922?answerId=733210022#733210022
